Converted zero-inflated distributions to Mixture subclasses
