+++ pymc3/tests/test_examples.py
@@ -253,7 +253,7 @@ class TestLatentOccupancy(SeededTest):
             # Estimated mean count
             theta = pm.Uniform('theta', 0, 100)
             # Poisson likelihood
-            pm.ZeroInflatedPoisson('y', theta, psi, observed=self.y)
+            pm.ZeroInflatedPoisson('y', psi, theta, observed=self.y)
         return model
 
     def test_run(self):
